refactor main handler
