+++ src/server.js
@@ -7,5 +7,5 @@ const port = process.env.PORT || 8080;
 const server = http.createServer(router)
 
 server.listen(port, () => {
-    console.log(`server is running at : http://localhost:${port}`);
+  console.log(`server is running at : http://localhost:${port}`);
 })
